Add files via upload
In this program, I utilized both Dynamic Memory Allocation and operator overloading to make a container that behaved similarly to a mathematical vector of arbitrary size. The program contains a class called DoubleContainer that consists a dynamically allocated array of double values. The container defines an Add, Clear, and Remove function, as well as overrides for numerous operators throughout the code.